Fail if no source file specified.
Change in behaviour. Before this fix the same directory
would be processed in fish if wildcard match fails.

Reason:
fish shell removes the argument if wildcard has no match.
bash passes the argument as it is.

e.g. try:
$ ./imgd -x 100x100 ~/dir/*.abc

With fish shell, the path is completely removed if there
is no file matching ~/dir/*.abc. Hence, we find no source
and try to process the current directory.